+++ merlin/models/tf/utils/batch.py
@@ -7,9 +7,9 @@ import tensorflow as tf
 from merlin.core.dispatch import DataFrameType, concat_columns, get_lib
 from merlin.schema import Schema, Tags
 
-from ...utils.schema import select_targets
-from ..core import Block, Model, RetrievalModel
-from ..dataset import Dataset
+from merlin.models.tf.core import Block, Model, RetrievalModel
+from merlin.models.tf.dataset import Dataset
+from merlin.models.utils.schema import select_targets
 
 
 class ModelEncode:
