+++ tests/tf/blocks/core/test_aggregation.py
@@ -22,6 +22,7 @@ import merlin.models.tf as ml
 from merlin.models.data.synthetic import SyntheticData
 from merlin.models.tf.blocks.aggregation import ElementWiseMultiply
 from merlin.schema import Tags
+from merlin.models.tf.blocks.core.aggregation import ElementWiseMultiply
 
 
 def test_concat_aggregation_yoochoose(testing_data: SyntheticData):
