+++ tests/tf/blocks/core/test_transformations.py
