+++ tests/tf/utils/test_batch.py
